Move classes to storage package
