+++ src/main/java/seedu/duke/CommandInterface.java
@@ -1,5 +1,7 @@
 package seedu.duke;
 
+import seedu.duke.storage.StorageInterface;
+
 public interface CommandInterface {
     void execute(TopicListInterface topics, StorageInterface storage);
