+++ src/test/java/seedu/duke/StorageTest.java
@@ -1,6 +1,7 @@
 package seedu.duke;
 
 import org.junit.jupiter.api.Test;
+import seedu.duke.storage.Storage;
 
 import java.util.ArrayList;
